requirements.txt
asgiref==3.7.2
astroid==2.15.6
black==23.7.0
click==8.1.7
dill==0.3.7
Django==4.2.4
django-utils-six==2.0
djangorestframework==3.14.0
install==1.3.5
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
mysqlclient==2.2.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pylint==2.17.5
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.7.1
wrapt==1.15.0
